backend/src/generate_token/requirements.txt
aws-xray-sdk==2.4.2
cryptography==2.7
pyjwt==1.7.1
